Add TextPhrase and TextPhraseOccurrence models, endpoints, and OpenAPI specifications
- Introduced `TextPhrase` and `TextPhraseOccurrence` models in the Prisma schema to manage multi-word phrase translations and their occurrences in texts.
- Added new API endpoints for retrieving phrase translations on specific text pages and managing phrases in the admin interface.
- Updated OpenAPI specifications to include the new endpoints and detailed response structures for phrase management.
- Enhanced documentation in `02-database.md`, `04-texts.md`, and `07-admin.md` to reflect the new features and their usage.
- Improved `TextService` and `AdminTextService` to support the new functionalities related to phrase occurrences and caching.

+++ src/markup-engine/unknown-word/unknown-word.processor.ts
@@ -90,7 +90,7 @@ export class UnknownWordProcessor {
       );
       await this.prisma.$executeRaw`
         UPDATE unknown_word uw
-        SET "seenCount" = uw."seenCount" + v.count,
+        SET "seenCount" = uw."seenCount" + v.count::integer,
             "lastSeen" = now(),
             "lastTextId" = COALESCE(${textId}, uw."lastTextId")
         FROM (VALUES ${values}) AS v(normalized, count)
